backend/requirements.txt
fastapi
uvicorn 
google-auth-oauthlib
google-auth
google-api-python-client
requests 